avm2: Store slot and disp ids for metadata tables as usizes

+++ core/src/avm2/globals/avmplus.rs
@@ -211,7 +211,7 @@ fn describe_internal_body<'gc>(
                     _ => unreachable!(),
                 };
 
-                let trait_metadata = vtable.get_metadata_for_slot(*slot_id);
+                let trait_metadata = vtable.get_metadata_for_slot(*slot_id as usize);
 
                 let variable = ScriptObject::new_object(activation);
                 variable.set_dynamic_property(istr!("name"), prop_name.into(), activation.gc());
@@ -274,7 +274,7 @@ fn describe_internal_body<'gc>(
 
                 let declared_by_name = declared_by.dollar_removed_name(mc).to_qualified_name(mc);
 
-                let trait_metadata = vtable.get_metadata_for_disp(*disp_id);
+                let trait_metadata = vtable.get_metadata_for_disp(*disp_id as usize);
 
                 let method_obj = ScriptObject::new_object(activation);
 
@@ -388,13 +388,13 @@ fn describe_internal_body<'gc>(
                 let metadata_object = ArrayObject::empty(activation);
 
                 if let Some(get_disp_id) = get {
-                    if let Some(metadata) = vtable.get_metadata_for_disp(*get_disp_id) {
+                    if let Some(metadata) = vtable.get_metadata_for_disp(*get_disp_id as usize) {
                         write_metadata(metadata_object, metadata, activation);
                     }
                 }
 
                 if let Some(set_disp_id) = set {
-                    if let Some(metadata) = vtable.get_metadata_for_disp(*set_disp_id) {
+                    if let Some(metadata) = vtable.get_metadata_for_disp(*set_disp_id as usize) {
                         write_metadata(metadata_object, metadata, activation);
                     }
                 }
